PrimsMaze code refactoring
