+++ MazeGenerator/Assets/Scenes/SampleScene.unity
@@ -531,7 +531,7 @@ GameObject:
   - component: {fileID: 787670778}
   - component: {fileID: 787670777}
   m_Layer: 0
-  m_Name: LevelBuilder
+  m_Name: Level
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -558,7 +558,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 787670776}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 699.10767, y: 306.49915, z: 9.757265}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
